Implemented the ProgressionVariable logic (using XpProgressor and XpData).
The developers add XP to the RpgCharacterData, which calls XpData to add XP. That XP is added internally and it’s checked whether there are enough XP to level up. If it determines such, it will use the assigned XpProgressor to calculate the new amount of XP to level up next time.